# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: XJU Lab API
  version: 1.0.0
  description: Session-authenticated API for the XJU Lab collaboration platform.
servers:
  - url: /api/v1
...
paths:
...
  /files:
    post:
      tags: [files]
      requestBody: { required: true, content: { multipart/form-data: { schema: { type: object, required: [file], properties: { file: { type: string, format: binary } } } } } }
      responses: { '201': { description: Validated private PDF, content: { application/json: { schema: { $ref: '#/components/schemas/FileObject' } } } }, '400': { $ref: '#/components/responses/ApiError' } }
  /files/{id}/content:
    get:
      tags: [files]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }, { name: Range, in: header, schema: { type: string, example: 'bytes=0-1023' } }]
      responses: { '200': { description: Private PDF content }, '206': { description: Single byte range }, '416': { description: Invalid or unsupported byte range }, '404': { $ref: '#/components/responses/ApiError' } }
  /files/{id}:
    delete:
      tags: [files]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { '204': { description: File marked for private object cleanup }, '404': { $ref: '#/components/responses/ApiError' } }
...
  /assessment/acm/imports:
    post:
      tags: [assessment, integrations]
      parameters: [{ name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 8, maxLength: 100 } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [sourceUrl], properties: { sourceUrl: { type: string, format: uri } } } } } }
      responses: { '202': { description: Full contest import accepted as an asynchronous job }, '400': { $ref: '#/components/responses/ApiError' }, '409': { $ref: '#/components/responses/ApiError' } }
  /assessment/acm/imports/{importId}:
    get:
      tags: [assessment, integrations]
      parameters: [{ name: importId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { '200': { description: Import state and safe result summary }, '404': { $ref: '#/components/responses/ApiError' } }
  /admin/members/{memberId}/oj-role-sync:
    get:
      tags: [members, integrations]
      parameters: [{ name: memberId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { '200': { description: Versioned external OJ role synchronization status } }
  /admin/members/{memberId}/oj-role-sync/retry:
    post:
      tags: [members, integrations]
      parameters: [{ name: memberId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { '200': { description: Latest desired role synchronization requeued }, '403': { $ref: '#/components/responses/ApiError' } }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
    printerAgentToken:
      type: http
      scheme: bearer
...
  schemas:
...
    FileObject:
      type: object
      required: [id, name, byteSize, pageCount, sha256]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        byteSize: { type: integer, format: int64 }
        pageCount: { type: integer }
        sha256: { type: string, pattern: '^[0-9a-f]{64}$' }
...
  responses:
    ApiError:
      description: Consistent request error
      content:
        application/json:
          schema:
            type: object
            required: [code, message, requestId]
            properties:
              code: { type: string }
              message: { type: string }
              requestId: { type: string }
              fieldErrors: { type: object, additionalProperties: { type: string } }
security:
  - sessionCookie: []